fix(anta): move main coroutine to a dedicated module and clean code
